requirements.txt
fastapi>=0.112,<0.116
python-telegram-bot>=21,<22
httpx>=0.27,<0.29
pg8000>=1.31,<2.0
pydantic>=2.6,<3.0
